webapp/requirements.txt
streamlit==1.44.1
langchain==0.3.23
langchain-huggingface==0.1.2
langchain-community==0.3.21
langchain-ollama==0.3.1
transformers==4.51.0
sentence-transformers==4.0.2
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
torch==2.6.0
chromadb
pymupdf4llm
accelerate==1.6.0
